add echo stmts to check existing files, edit comments

+++ postMEP_volumetric_analysis/create_vol_parc_prep.sh
@@ -17,6 +17,8 @@ for i in $(cat "$SubjectsList"); do
     # Make dir for symlinks if necessary
     if [ ! -d "$SubFreeSurferDir/symlinks" ]; then
         mkdir "$SubFreeSurferDir/symlinks"
+    else
+        echo -e "\n$SubFreeSurferDir/symlinks already exists...\n"
     fi
 
     # Replace Left Pial symbolic link with file it points to
@@ -24,6 +26,8 @@ for i in $(cat "$SubjectsList"); do
         mv "$SubSurfDir/lh.pial" "$SubFreeSurferDir/symlinks"
         cp "$SubSurfDir/lh.pial.rawavg.conf" "$SubFreeSurferDir"
         mv "$SubFreeSurferDir/lh.pial.rawavg.conf" "$SubSurfDir/lh.pial"
+    else
+        echo -e "$SUBJ lh.pial already exists." # check
     fi
 
     # Replace Left White symbolic link with file it points to
@@ -32,7 +36,7 @@ for i in $(cat "$SubjectsList"); do
         cp "$SubSurfDir/lh.white.rawavg.conf" "$SubFreeSurferDir"
         mv "$SubFreeSurferDir/lh.white.rawavg.conf" "$SubSurfDir/lh.white"
     else
-        echo -e "$SUBJ lh.white already exists." # to check code is working
+        echo -e "$SUBJ lh.white already exists." # check
     fi
 
     # Replace Left White.H symbolic link with file it points to
@@ -40,6 +44,8 @@ for i in $(cat "$SubjectsList"); do
         mv "$SubSurfDir/lh.white.H" "$SubFreeSurferDir/symlinks"
         cp "$SubSurfDir/lh.white.preaparc.H" "$SubFreeSurferDir"
         mv "$SubFreeSurferDir/lh.white.preaparc.H" "$SubSurfDir/lh.white.H"
+    else
+        echo -e "$SUBJ lh.white.H already exists." # check
     fi
 
     # Replace Left White.K symbolic link with file it points to
@@ -47,6 +53,8 @@ for i in $(cat "$SubjectsList"); do
         mv "$SubSurfDir/lh.white.K" "$SubFreeSurferDir/symlinks"
         cp "$SubSurfDir/lh.white.preaparc.K" "$SubFreeSurferDir"
         mv "$SubFreeSurferDir/lh.white.preaparc.K" "$SubSurfDir/lh.white.K"
+    else
+        echo -e "$SUBJ lh.white.K already exists." # check
     fi
 
     # Replace Right Pial symbolic link with file it points to
@@ -54,6 +62,8 @@ for i in $(cat "$SubjectsList"); do
         mv "$SubSurfDir/rh.pial" "$SubFreeSurferDir/symlinks"
         cp "$SubSurfDir/rh.pial.rawavg.conf" "$SubFreeSurferDir"
         mv "$SubFreeSurferDir/rh.pial.rawavg.conf" "$SubSurfDir/rh.pial"
+    else
+        echo -e "$SUBJ rh.pial already exists." # check
     fi
 
     # Replace Right White symbolic link with file it points to
@@ -61,6 +71,8 @@ for i in $(cat "$SubjectsList"); do
         mv "$SubSurfDir/rh.white" "$SubFreeSurferDir/symlinks"
         cp "$SubSurfDir/rh.white.rawavg.conf" "$SubFreeSurferDir"
         mv "$SubFreeSurferDir/rh.white.rawavg.conf" "$SubSurfDir/rh.white"
+    else
+        echo -e "$SUBJ rh.white already exists." # check
     fi
 
     # Replace Right White.H symbolic link with file it points to
@@ -68,6 +80,8 @@ for i in $(cat "$SubjectsList"); do
         mv "$SubSurfDir/rh.white.H" "$SubFreeSurferDir/symlinks"
         cp "$SubSurfDir/rh.white.preaparc.H" "$SubFreeSurferDir"
         mv "$SubFreeSurferDir/rh.white.preaparc.H" "$SubSurfDir/rh.white.H"
+    else
+        echo -e "$SUBJ rh.white.H already exists." # check
     fi
 
     # Replace Right White.K symbolic link with file it points to
@@ -75,6 +89,8 @@ for i in $(cat "$SubjectsList"); do
         mv "$SubSurfDir/rh.white.K" "$SubFreeSurferDir/symlinks"
         cp "$SubSurfDir/rh.white.preaparc.K" "$SubFreeSurferDir"
         mv "$SubFreeSurferDir/rh.white.preaparc.K" "$SubSurfDir/rh.white.K"
+    else
+        echo -e "$SUBJ rh.white.K already exists." # check
     fi
 
     # Make temporary copies of FreeSurfer subdirs in main subject dir
@@ -82,9 +98,12 @@ for i in $(cat "$SubjectsList"); do
         cp -rf "$SubFreeSurferDir" "$SUBJECTS_DIR/$SUBJ"
     fi
 
-    mv "$SUBJECTS_DIR"/"$SUBJ"/"$SUBJ"/* "$SUBJECTS_DIR/$SUBJ" 
-    rm -r "$SUBJECTS_DIR/$SUBJ/$SUBJ"
+    # Move files into main subject dir and delete extra dir
+    if [ -d "$SUBJECTS_DIR/$SUBJ/$SUBJ" ]; then
+        mv "$SUBJECTS_DIR"/"$SUBJ"/"$SUBJ"/* "$SUBJECTS_DIR/$SUBJ" 
+        rm -r "$SUBJECTS_DIR/$SUBJ/$SUBJ"
+    fi
 
 done
 
-echo "\n    ---------------------    All Subjects Done    ---------------------    \n"
+echo "\n    ---------------------    Done    ---------------------    \n"
